% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{startup}
\alias{startup}
\title{Start up the DNABarcodeCompatibilityShiny interface}
\usage{
startup()
}
\value{
Launches the web interface
}
\description{
Start up the DNABarcodeCompatibilityShiny interface
}
\examples{
# Just run startup()
ls()

}
